+++ packages/golden-layout/index.d.ts
@@ -415,7 +415,7 @@ declare module '@glue42/golden-layout' {
          * Controls the visibility of the close button in the window tabs
          */
             showCloseButton?: boolean;
-
+          allowDropHeader?:boolean;
             /**
           * Enables or disables the abilitiy to drop windows in the left area of the workspace
           */