+++ packages/workspaces-ui-core/changelog.md
@@ -1,4 +1,5 @@
 1.4.5
+feat: implemented allowSplitters to rows and columns and allowDropLeft, allowDropTop, allowDropRight, allowDropBottom, allowDropHeader to the groups and added style improvements
 fix: started firing open and close workspace events when the last workspace in a frame acting like a platform has been closed
 1.4.4
 chore: bump due to dependencies update